mostly done updating svg text, better for FF & mathjax
FF doesn't yet support baseline-shift, so put a (uber-hacky)
workaround into Plotly.util.convertToTspans.

Then it became clear that Plotly.Lib.styleText wasn't amenable
to the same hackery, so I swapped the remaining things using it
over to convertToTspans

That brought mathjax support to various places that were missing
it, and getting that to work right is a royal pain. So far I
think it's all working except hovertext