Plot as synchronously as possible to minimize repaints
Plotly.Lib.syncOrAsync: take a sequence of function calls, each of
which can return a promise if it does something asynchronous, or
something else if it's synchronous. Execute this sequence as
synchronously as possible.

Still need to integrate this into restyle and relayout, though
it's less important there than in plot