fix installation from npm3
npm3 will flatten node_modules, so the nested node_modules that is used
for copying the JSON over from Topojson will not be available in the
package's relative node_modules. This commit uses a synchronous check to
see if the path is valid and on failure with try to use the path
expected from npm3 installations.