Add the ability to use aggregate values for box plots
Add a new calcdataOverride property for box plot objects which will
allow calculated box plot values to be replaced. To use, pass the
aggregate values as the main data along with calcdataOverride property.
Plotly will calculate all of the values so it is only necessary to
override some of the values. Look at the calculateStats function in
Plotly.js to see the properties that can be overridden.