Sort out interaction between drag/edit and legend toggles.
Introduces a global _editing flag to be able to block other interactions while edit is in progress. Could be one solution for #1437.

This makes it more clear that there should be some refactor to consolidate the double click behavior, since when the drag layer is on, that code path's click detection is used. The other path needs to remain since the drag layer doesn't exist when editable: false, but it's now easier to picture a world where these two paths converge.