Fix RE used to correct quotes used in URLs for IE.
The `svgToImg()` function tries to normalize the SVG for compatibility
with IE.  However, currently it use a regular expression that is too
greedy when matching the attribute values that need quotes changed from
single to double.  The result is that invalid XML is generated and will
cause IE11 to fail with a `XML 5607` error and `svgToImg()` will never
return.  Edge will not complain but will receive invalid XML, which
causes problems when used.

This change restricts matching to only data surrounded by single quotes.

Seems like a jasmine test should be added for `svgToImg()` along the
lines of

  `it('Returns valid XML',...)`