Added support for encoding TypedArrays as primitive objects (representation objects)
A TypedArray representation object has two properties: `dtype` and `data`.

`dtype` is a string indicating the type of the typed array (`'int8'`, `'float32'`, `'uint16'`, etc.)

`data` is a primitive JavaScript object that stores the typed array data. It can be one of:
 - Standard JavaScript Array
 - ArrayBuffer
 - DataView
 - A base64 encoded string

The representation objects may stand in for TypedArrays in `data_type` properties and in properties with `arrayOk: true`.

The representation object is stored in `data`/`layout`, while the converted TypedArray is stored in `_fullData`/`_fullLayout`