Support rendering tex even when global MathJax rendering mode is not SVG
 1. Specify the SVG font
 2. Set rendering mode to SVG just prior to typesetting, then restore it to it's original value just afterwards

With this change, latex rendering works in the following MathJax configurations:
 - TeX-MML-AM_CHTML
 - TeX-MML-AM_HTMLorMML
 - TeX-MML-AM_SVG
 - TeX-AMS-MML_HTMLorMML
 - TeX-AMS_CHTML
 - TeX-AMS_SVG
 - TeX-AMS_HTML
 - TeX-AMS-MML_SVG

(It still does not work if the global MathJax TeX extensions are not loaded.)

Previously it only worked in the TeX-AMS-MML_SVG configuration. Also, with these changes latex
rendering works without any additional configuration when used from Python (and probably R) inside
the Jupyter Notebook.