Create tmpDiv on the `Queue` just before Typesetting
This avoids race condition where an external typesetting operation
is in progress (or already on the Queue) at the start `texToSVG`
function.

With this change, we no longer need to set `skipStartupTypeset` to true,
which alters the global default MathJax behavior of typesetting the full page
automatically on startup. Other utilities, such as nbconvert, have
not worked properly with plotly because they rely on this default
behavior