Add support for a new global window.PlotlyConfig configuration object
Add a MathJaxConfig option that, when set to 'local', causes plotly.js
to bypass all global MathJax configuration. Regardless of the value of
MathJaxConfig, all MathJax rendering commands are surrounded by logic
to properly configure the MathJax properties that plotly.js needs.

When MathJaxConfig is not 'local', plotly.js will continue to setup
the global MathJax config as it always has. This is the default behavior
for backward compatibility.  When plotly.js is used in a situation
where it is not the only component using MathJax, then MathJaxConfig
should be set to 'local' to keep from overwriting the user's desired
global configuration.