First iteration of adding " domain" to x and y ref specifiers
This allows shapes and eventually annotations to refer to the length of
an axis, so if this length is changed, the shape changes accordingly.
For example for a shape, setting xref: "x2 domain", x0: 0, x1: 1 makes a
shape that spans exactly the width of xaxis2. When this axis changes
length (via the setting of its 'domain' attribute), then the shape will
span the new axis.