cleanId accepts domain referencing IDs
But the acceptance of domain referencing IDs is only enabled for certain
elements, namely images, annotations, and shapes. So cleanAxRef has been
adapted to pass the domain argument to cleanId and accept such elements.
cleanAxRef is what cleans the IDs on these kinds of elements.

Furthermore, cleanAxRef is called on image elements now in
src/plot_api/helpers.js.