Annotation arrows can be placed like in data coordinates
This works now with 'paper' and 'domain' axis reference types.
However, the annotations are not yet properly moveable in with 'paper'
axis reference types. Note that the placement is absolute, not relative
as when 'axref' or 'ayref' are set to 'pixel'. This is consistent with
how these already work when the arrow's axis reference is set to data
coordinates.

For this to work, the arrow's xref and yref (i.e., axref, ayref) must be
set to exactly the same thing as the annotation's xref and yref (xref,
yref), otherwise arrow's corresponding axis reference will fall back to
'pixel' and be specified relative to the annotation's position. This is
constent with how this already works with data coordinates.