Working on the annotation tests
It seems that "paper" doesn't work as expected when axref or ayref are
set as the same as xref and or yref (who are also set as "paper").

Indeed, the "range" and "domain" reference types do work the new way,
where the arrow head can be placed absolutely as long as axref or ayref
is exactly the same as the corresponding xref or yref.

So "paper" should take up this functionality, or the spec should make it
clear that it doesn't work and the test updated accordingly.