Static shape, annotation and image tests are passing
What seemed like the problem was that the gd.layout remained referring
to a single object obtained from the
test/image/mocks/domain_ref_base.json file, so if we changed the layout
in one test, it remained changed in the other test. What this seems to
mean is that Plotly.newPlot doesn't make a copy of the layout it's
handed (?). Any how, this was fixed by calling `Lib.extendDeep({},mock)`
for each test, which obtains a new object pointing to new memory and
therefore will not be affected by previous tests.

Tests for editable mode do not exist yet.