Numeric categories in SVG paths intepreted as indices
This is because SVG paths are always strings, so there is no practical
way of indicating a number was meant and not a numeric string when
specifying SVG paths (sure they could be wrapped in escaped quotes but
not today). So when a number is encountered in an SVG (i.e., it is a
numeric string), then it is converted to a number so that it is
interpreted as an index into layout.categoriesarray.