Add willReadFrequently to 2d canvas contexts, where appropriate.
This chang add the willReadFrequently:true 2d context creation
attribute to call sites where canvases will be read from (i.e
getImagData or toDataURL).  This is a performance hint that tells
the browser to optimize for readbacks.

See: https://html.spec.whatwg.org/multipage/canvas.html#concept-canvas-will-read-frequently

The new API will soon be launched in chromium-based web browsers.

The objective of this change is to optimize performance and prevent
possible performance regressions when this new feature is launched
by major web browser in the near future.