Added static gradient fills for scatter traces.
Adds a new attribute `fillgradient` to `scatter` traces
which allows the user to specify a gradient as a colorscale
and an orientation in which it will be applied. `fillgradient`
also has optional start and stop attributes which can be
used to define absolute start and stop points in plot coordiantes.
This allows to equip different traces with exactly matching fill gradients.
Omitting start and stop will apply the gradient between the extrema
of the fill polygon (along the fill direction).

When a legend is displayed, multiple passes are made through the
`Drawing.setFillStyle` function, for the legend previews and the
actual plot, during which the axes are re-scaled.
This leads to errors with gradients not computing start and stop
coordinates correctly. To work around this, `setFillStyle` was
modified to allow distinguishing between legend and plot passes.