Intermediate commit w/ partial work- new strat:
So, parameterization is good in testing.

Promises + Jasmine + Karma w/o let or const make it hard to do manually.

There should be a parameterization library.

There is a way to make it all work (moving all loops outside of
it/describe).

Which is good, because then each parameter can be reported as a separate test,
which improves reporting. But that solution is generally unpopular.