Display hover label at cursor position as fallback for scatter traces.
For curved edges of fills there is a chance that the hover detection polygons miss a point and the correct hover label position cannot
be determined. Previous code fell back to positioning the label
at the largest edge of any polygon of the trace of their combined
vertical midpoint, with undetermined behaviour if no polygon
overlapped the midpoint.
This change instead falls back to simply displaying the label at the
current cursor position, which simplifies code and results in less
confusing label placement.