Add config so dev can place closed captions:
Adding a closed caption div to plotly interefered with plotly's
management of it.

We now start by default outside of the generated plotly divs.

But we add options to users can a) generate it with any id/class.

Or specify that it already exists and use an existing element by id.