Improve numerical stability of projection
When the position vector has large values that are cancelled out
by large values in the model matrix, it is more numerically stable
to first multiply the position by the model matrix instead of
multiplying the matrices together first.

Use forks of the gl-vis libraries that use the problematic pattern.