Fix unit tests and bug applying bgcolor with relayout
Fixed unit tests by updating the expectations as appropriate to match
the changed logic in Pull Request #7109. This revealed a bug where
background color changes applied using `Plotly.relayout()` were missed.
This bug was also fixed in this change in order to get all unit tests,
as updated, to pass.